Avoid finding kernel modules multiple times
Revert the first hunk from

 commit 05fb94d2c42abe9cfd86c3663d704c268f325503
 Author: Daniel P. Berrange <berrange@redhat.com>
 Date:   Wed Apr 4 16:30:49 2012 +0100

    Search kernel module dirs & fix mem leaks

The change to remove '/kernel' from the module directory
was bogus. The real fix was in the Fedora kernel RPM
packaging

Without the '/kernel' suffix, the same module might be
found multiple times. eg When /lib/modules/`uname -r`/
has symlinks to the compiled sources (through /build and
/source) it crashed virt-sandbox with the error:

   "Unable to start sandbox: Error opening file
   '/tmp/libvirt-sandbox-9ivpRN/9pnet.ko': File exists"