Typo and example fix
I fixed a typo in the strace debug feature, if you specified
LIBVIRT_SANDBOX_STRACE=poll it would write "strace =poll" in the
kernel command line and consequently it wouldn't get picked up because
of that extra space.
Also, the example virt-sandbox.py was a bit outdated so I updated it.
The shell.py doesn't work as it has serial1 hardcoded for qemu. I
haven't checked the rest.

commit c9258ea3485a20c0b02f261fd9b8de4af32bf201
Author: Radu Caragea <sinaelgl@gmail.com>
Date:   Tue Jun 19 12:18:02 2012 +0300

    Fix python example and typo in strace kernel cmdline