Maintain only one list of all mount types
Currently there are seprate lists maintained for host bind
mounts, host image mounts and guest bind mounts. This means
that mounts can't be processed in the order the user
requested, which means a host bind mount can't be made on
top of a host image mount.

Switch to only having one single list of mounts in the
config API