better wording when connecting to a console patch from  Michael Scherer.
"Escape character is '^]'." is the message show by telnet as well as the one of virt-sandbox-service when connecting to the console of a service. But the keyboard shortcut doesn't do the same thing ( ie, on telnet, you have a prompt ).

While telnet is likely to be unheard from newer admins, I think a better message could be :
Type 'Ctrl + ]' to detach from the console.