Add ability to directly open console pty device paths
Using a libvirt stream object for accessing the console is good
for privilege separated environments, or remote connections, but
it requires that each sandbox hold open a libvirt connection
for its lifetime. This quickly hits the libvirt connection limit
of 20.

Add an option to directly open the PTY device associated with
the console, instead of using the stream object. This will (later)
allow the connection to be closed while the sandbox is running.