Rename tmpdir to statedir in sandbox builder classes
The directory where we store sandbox state files is not really
a temporary directory, but rather a state directory, which is
cleaned up based on some arbitrary rules. Rename the parameter
pass to the sandbox builders from 'tmpdir' to 'statedir' to
reflect this. Also stop explicitly passing around 'configdir'
since this can be derived from 'statedir' when needed

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>