Stop using broken shutil.copytree() method
The shutil.copytree() method is broken in several ways

 - Raises exceptions if it sees a socket or fifo
 - Copies block/char device content into plain files
 - Calls stat() far too many times

Fixing this requires passing a callback to filter the
file list, which requires more use of stat() making it
even less efficient.

Ditch it and write a method that works correctly for
our needs, skipping block/char/fifo/socket files entirely.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>