Fix symlink path in multi-user.target.wants
Sinc we switched to using a drop-in unit file override in
change 775cebb4e2d5cd04d9c4ceb0a48dd36e22a3682c, the symlink
from multi-user.target.wants is broken. We must link to the
original source unit file via the fully qualified path.

Systemd didn't actually mind that the symlink was dangling,
but it is safer to ensure it is correct, in case systemd gets
more strict in the future.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>