configure: Require newer libvirt-glib
In the commit 68406aff new feature was introduced - users are allowed
to select the format of mounted host images. However, this uses some
libvirt-glib features that were introduced in the 0.2.1 release, e.g.
the symbol GVIR_CONFIG_DOMAIN_FILESYS_DRIVER_LOOP was introduced in
dbd063fe. So we must update the minimal required version of
libvirt-glib. At the same time, update it in the spec file too.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>