Make init-lxc all static
When running a sandbox with a / different from the host one, we will
need to copy all init-lxc dependencies into a mounted folder... but we
have no way to tell libvirt to set the LD_LIBRARY_PATH for the init
command.

Turning init-lxc all-static help us work around that problem, and drops
the useless dependencies on glib and libvirt-sandbox.