virt-sandbox-image: switch to use URI to identify templates
Currently the CLI syntax is somewhat docker specific requiring
inclusion of --registry arg to identify the docker download
server. Other app containers have a notion of download server,
but don't separate it from the template name.

This patch removes that docker-ism by changing to use a URI
for identifying the template image. So instead of

  virt-sandbox-image download \
      --source docker --registry index.docker.io
      --username dan --password 123456 ubuntu:15.04

You can use

  virt-sandbox-image download docker://dan:123456@index.docker.io/ubuntu?tag=15.04

The only mandatory part is the source prefix and image name, so
that can shorten to just

  virt-sandbox-image download docker:///ubuntu

to pull down the latest ubuntu image, from the default registry
using no authentication.