docker: introduce a DockerImage class
Instead of directly using the Template class in docker code,
introduce a DockerImage class which represents the data in a
more convenient manner. In particular it will canonicalize
the image names eg a bare image name with no "/", like "ubuntu",
should be "library/ubuntu"

The index.json file now stores the repo name, image name and tag
so we resolve the correct matching image where there are multiple
versions of the same image stored locally.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>