Pass verbose values to init-common
libvirt-sandbox-init-common is expecting -v parameter to set it in
verbose mode... but it will never be passed by the launcher program.
Writing the core.verbose parameter in the sandbox configuration
makes that values actually usable from the init.

While at it, add and set a core.debug value to the sandbox configuration
to allow setting only the common init in debug mode, not the
virtualization specific one.