Fix argparser incompatibilities with newer python 3
Python 3 changes such that if no subparser command is listed, it just
throws an error. To get back the old behavior we need to set the
'required'  attribute and a dest name.

Since we treat 'debug' as a global attribute we need to set that on the
top level parser too, otherwise we get a missing attribute error with
newish python 3 too.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>