gitlab: introduce CI jobs testing git master & distro libvirt
The sandbox build needs to validate one axis

  - A variety of libvirt versions

We test a variety of libvirt versions by running a build against the
distro provided libvirt packages. All that is then missing is a build
against the latest libvirt git master, which only needs to be run on
a single distro, for which Ubuntu 20.04 is picked as a stable long life
base.

Several distros are not included because they don't provide the package
pre-requisites.

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>